JAVA-3295: unify transaction and session tests
