+++ driver-core/src/main/com/mongodb/internal/connection/DefaultServer.java
@@ -159,6 +159,15 @@ class DefaultServer implements ClusterableServer {
     }
 
     @Override
+    public void invalidate(final Throwable t, final SessionContext sessionContext) {
+        notNull("sessionContext", sessionContext);
+        invalidate(t);
+        if (t instanceof MongoSocketException && sessionContext.hasSession()) {
+            sessionContext.markSessionDirty();
+        }
+    }
+
+    @Override
     public void close() {
         if (!isClosed()) {
             connectionPool.close();
@@ -220,10 +229,7 @@ class DefaultServer implements ClusterableServer {
                 invalidate();
                 return (T) e.getResponse();
             } catch (MongoException e) {
-                if (e instanceof MongoSocketException && sessionContext.hasSession()) {
-                    sessionContext.markSessionDirty();
-                }
-                invalidate(e);
+                invalidate(e, sessionContext);
                 throw e;
             }
         }
@@ -241,10 +247,7 @@ class DefaultServer implements ClusterableServer {
                             invalidate();
                             callback.onResult((T) ((MongoWriteConcernWithResponseException) t).getResponse(), null);
                         } else {
-                            if (t instanceof MongoSocketException && sessionContext.hasSession()) {
-                                sessionContext.markSessionDirty();
-                            }
-                            invalidate(t);
+                            invalidate(t, sessionContext);
                             callback.onResult(null, t);
                         }
                     } else {
