+++ driver-core/src/test/unit/com/mongodb/internal/connection/TestServer.java
@@ -25,6 +25,7 @@ import com.mongodb.connection.ServerDescription;
 import com.mongodb.connection.ServerId;
 import com.mongodb.event.ServerDescriptionChangedEvent;
 import com.mongodb.event.ServerListener;
+import com.mongodb.session.SessionContext;
 
 import static com.mongodb.connection.ServerConnectionState.CONNECTING;
 
@@ -57,6 +58,11 @@ public class TestServer implements ClusterableServer {
 
     @Override
     public void invalidate(final Throwable reason) {
+        invalidate();
+    }
+
+    @Override
+    public void invalidate(final Throwable reason, final SessionContext sessionContext) {
         invalidate();
     }
